app: add icons to voucher delete

+++ app/src/app/campaigns.module/components/campaign-info.component/campaign-voter-input.component/campaign-voter-input.component.ts
@@ -98,6 +98,9 @@ export class CampaignVoterInputComponent implements OnInit {
       return;
     }
 
+    if (!confirm('Delete this voucher?')) {
+      return;
+    }
 
     await this.vouchersService.delete(entry.voucherId);
